backend/requirements.txt
Django 
python-decouple
whitenoise